# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger for the API that was developed as a core platform of the BUILDSPACE project.\n**Note** that the **X-Group-Id** is a header that is required in all the enpoints.",
        "title": "BUILSPACE Core Platform Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "BUILDSPACE Core Platform Support",
            "url": "http://www.swagger.io/support",
            "email": "isotiropoulos@singularlogic.eu"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/file": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This is the endopoint to update file meta data. Pass a models.File of the file that will be updated with the updates included.\n**Note** that this endpoint updates the meta data and not the file contents. To update file contents user must delete int and re-upload it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Update a file.",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.File"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.File"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This is the endopoint to upload files. The files are uploaded using a multipart streaming upload.\nStep 1 is to select the content-type.\n- If **application/json** then the request will be sent to initialize the multipart upload. In this case user must pass a **File model as a payload** containing the **folder** and the **original_title** fields. User must also pass the **total** header to specify the number of parts that will be uploaded.\n- If **application/octet-stream** user must pass the **binary data** (decoded) in the body and also provide the **file ID** and part number parameters.",
                "consumes": [
                    "application/json",
                    "application/octet-stream"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Upload a file.",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Total parts of multipart upload",
                        "name": "total",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "file",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Number of part",
                        "name": "part",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.File"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            }
        },
...
        "/info/file": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the metadata of a file by it's ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Get metadata of file.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorReport": {
            "type": "object",
            "properties": {
                "internal_status": {
                    "description": "Status of the error",
                    "type": "string"
                },
                "message": {
                    "description": "Message of the error",
                    "type": "string"
                },
                "reason": {
                    "description": "Reason of the error",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the error",
                    "type": "integer"
                }
            }
        },
        "models.File": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "File's id",
                    "type": "string"
                },
                "ancestors": {
                    "description": "All ancestor folders",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "file_type": {
                    "description": "The file's extention",
                    "type": "string"
                },
                "folder": {
                    "description": "Parent folder of the file",
                    "type": "string"
                },
                "meta": {
                    "description": "File's Metadata",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Meta"
                        }
                    ]
                },
                "original_title": {
                    "description": "The file's title before uploading",
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}